Разбиение классов на файлы
